+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Aggregation Request Types
+
+### Fixed
+- Added missing `keyword` type
 
 ## [0.1.5] - 2020-08-12
 ### Fixed